validation if user not found in delete user

+++ app/Http/Controllers/UserController.php
@@ -15,8 +15,10 @@ class UserController extends Controller
     public function destory(Request $request, $id) {
         $user = User::where('id', $id)->first();
 
+        if (empty($user)) return $this->errorResponse('user not found', [], 404);
+
         if ($user->isAdmin()) {
-            return $this->successResponse('can\'t delete admin', [], 403);
+            return $this->errorResponse('can\'t delete admin', [], 403);
         }
 
         $user->delete();
